requirements.txt
amqp==2.4.2
billiard==3.5.0.5
celery==4.2.1
certifi==2018.11.29
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.1.4
django-cors-headers==2.4.0
django-mail-queue==3.1.6
django-money==0.14.4
django-rest-swagger==2.1.2
djangorestframework==3.9.0
idna==2.8
itypes==1.1.0
Jinja2==2.10
kombu==4.4.0
MarkupSafe==1.1.0
mysqlclient==1.3.14
openapi-codec==1.3.2
py-moneyed==0.8.0
pytz==2018.7
requests==2.21.0
simplejson==3.16.0
uritemplate==3.0.0
urllib3==1.24.1
vine==1.2.0
